feat: add upstream info into the accesslog

+++ pkg/router/accesslog/upstream_test.go
@@ -13,7 +13,7 @@ func TestWithUpstream(t *testing.T) {
 	u := &Upstream{
 		Addr:           "http://127.0.0.1:8080/fission-function/fn-1",
 		ResponseLength: 100,
-		ResponseTime:   int(time.Second),
+		ResponseTime:   int64(time.Second),
 		ResponseStatus: http.StatusOK,
 	}
 	ctx := WithUpstream(context.Background(), u)
